Add ActionSignalDispatcher utility class for managing multiple signals with the same tick counter.
NES: Migrate PPU and APU classes to use the new ActionSignalDispatcher utility

+++ core/src/main/java/io/github/arkosammy12/jemu/core/util/ActionSignal.java
@@ -19,7 +19,7 @@ public final class ActionSignal {
     }
 
     public void trigger(int delay, int value) {
-        long fireAt = this.ticks + delay;
+        long fireAt = this.ticks + (long) delay;
         this.timers.enqueue(fireAt);
         this.pendingValues.put(fireAt, value);
     }
